Refactor git-integration entry points
Organize into subpackage in order to keep main folder clean, and better
accomodate other VCS integrations.
