+++ nbdime/vcs/__init__.py
@@ -1,0 +1,1 @@
+"""Contains entry points for VCS integration."""
