+++ nbdime/vcs/git/__init__.py
@@ -1,0 +1,1 @@
+"""Contains entry points for git integration."""
